+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to laravel-userstamps will be documented in this file.
 
+### 0.0.4
+- Added the column type of the id of the users table to the config. This was 
+needed to make sure the foreign key could be added to the table where the
+column types are matching.
+
 ### 0.0.3
 - Renamed the trait for easier reading